[Refactor] 패키지 구조 변경
기능에 변경이 생기는 경우 파악하기 쉽게 만들기 위해 기능 중심으로 패키지를 변경
