+++ app/src/main/java/jtot/dev/feature/main/MainActivity.kt
@@ -1,7 +1,8 @@
-package jtot.dev
+package jtot.dev.feature.main
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import jtot.dev.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
