+++ app/src/main/java/jtot/dev/feature/play/decoration/ContentDecoration.kt
@@ -1,4 +1,4 @@
-package jtot.dev.decoration
+package jtot.dev.feature.play.decoration
 
 import android.graphics.Rect
 import android.view.View
